Possible fixes to consistency. Written more stress-testing code under defines (off by default).
Fixed colliders for the character for various stances.
Fixed collider generation - the rotation didn't take normal vectors into
consideration.

Using IE for floats, II for ints in randomization.

Using floats only for delta - fixes padding issues.

Serialization appears to work stably, but we still have desyncs due to
what appears to be triggered by round restarts.

+++ src/application/arena/arena_handle.h
@@ -105,7 +105,7 @@ public:
 				return 1.0;
 			}
 			else {
-				const auto current_logic_speed = m.round_speeds.logic_speed_mult;
+				const auto current_logic_speed = static_cast<double>(m.round_speeds.logic_speed_mult);
 				const auto chosen_audiovisual_speed = rules.view.audiovisual_speed;
 
 				return std::max(current_logic_speed, chosen_audiovisual_speed);
